+++ template/RELEASE_NOTES.md
@@ -12,6 +12,7 @@
 * ANDDEP-459 Added transition to application settings on DebugScreen
 * Fixed debug screen push notification appearing on each App#onCreate invocation
 * ANDDEP-770 Fixed textStyle overriding and proguard bugs
+* ANDDEP-599 Added Completable check to ResponseTypeAdapterFactory to avoid parse exceptions
 
 ## 0.4.0
 ##### Template